Refactor the activities boxes

+++ lib/screens/rubric_screen.dart
@@ -100,7 +100,7 @@ class _RubricScreenState extends State<RubricScreen> {
                                                 rubricScreenCubit.setScore(
                                                     stateRubric.activity,
                                                     index,
-                                                    i,
+                                                    i.toDouble(),
                                                     presenter);
                                               },
                                             ),
